+++ mlxtend/feature_extraction/__init__.py
@@ -1,0 +1,10 @@
+# Sebastian Raschka 2014-2016
+# mlxtend Machine Learning Library Extensions
+# Author: Sebastian Raschka <sebastianraschka.com>
+#
+# License: BSD 3 clause
+
+from .principal_component_analysis import PrincipalComponentAnalysis
+
+
+__all__ = ["PrincipalComponentAnalysis"]